requirements.txt
neo4j>=5.18.0

google-generativeai>=0.4.1
pydantic>=2.6.0

python-dotenv>=1.0.1